Fixed issues with the links in controller specs and wound back logging changes
