+++ conf/application.conf
@@ -22,7 +22,6 @@ appName=company-registration
 play.modules.enabled += "com.kenshoo.play.metrics.PlayModule"
 play.modules.enabled += "play.modules.reactivemongo.ReactiveMongoHmrcModule"
 play.modules.enabled += "config.Module"
-play.modules.enabled += "submissionModules.HeldSubmissionModule"
 
 #Required for object based play2.5 applications
 play.http.requestHandler = "play.api.http.GlobalSettingsHttpRequestHandler"
