+++ test/controllers/AccountingDetailsControllerSpec.scala
@@ -20,14 +20,14 @@ package controllers
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import fixtures.{AccountingDetailsFixture, AuthFixture}
-import helpers.SCRSSpec
+import helpers.{ControllerHelper, SCRSSpec}
 import mocks.{MockMetricsService, SCRSMocks}
 import models.{AccountPrepDetails, ErrorResponse}
 import org.mockito.Matchers
 import org.mockito.Mockito._
 import org.scalatest.mock.MockitoSugar
 import org.scalatestplus.play.{OneAppPerSuite, OneAppPerTest}
-import play.api.libs.json.Json
+import play.api.libs.json.{JsObject, Json}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import services.PrepareAccountService
@@ -35,7 +35,7 @@ import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.Future
 
-class AccountingDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSMocks with  AuthFixture with AccountingDetailsFixture {
+class AccountingDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSMocks with  AuthFixture with AccountingDetailsFixture with ControllerHelper{
 
   implicit val system = ActorSystem("CR")
   implicit val materializer = ActorMaterializer()
@@ -66,7 +66,10 @@ class AccountingDetailsControllerSpec extends UnitSpec with MockitoSugar with SC
 
       val result = controller.retrieveAccountingDetails(registrationID)(FakeRequest())
       status(result) shouldBe OK
-      await(jsonBodyOf(result)) shouldBe Json.toJson(validAccountingDetailsResponse)
+
+      val json =  await(jsonBodyOf(result)).as[JsObject]
+      val links2 = links(json.value("links").as[JsObject])
+      json.as[JsObject] - "links" ++ links2 shouldBe Json.toJson(Some(validAccountingDetailsResponse))
     }
 
 
@@ -126,7 +129,10 @@ class AccountingDetailsControllerSpec extends UnitSpec with MockitoSugar with SC
 
       val result = call(controller.updateAccountingDetails(registrationID), response)
       status(result) shouldBe OK
-      await(jsonBodyOf(result)) shouldBe Json.toJson(validAccountingDetailsResponse)
+
+      val json =  await(jsonBodyOf(result)).as[JsObject]
+      val links2 = links(json.value("links").as[JsObject])
+      json.as[JsObject] - "links" ++ links2 shouldBe Json.toJson(Some(validAccountingDetailsResponse))
     }
   }
