+++ test/controllers/CompanyDetailsControllerSpec.scala
@@ -20,10 +20,10 @@ import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import connectors.AuthConnector
 import fixtures.{AuthFixture, CompanyDetailsFixture}
-import helpers.SCRSSpec
+import helpers.{ControllerHelper, SCRSSpec}
 import org.mockito.Matchers
 import org.mockito.Mockito._
-import play.api.libs.json.Json
+import play.api.libs.json.{JsObject, Json}
 import play.api.test.FakeRequest
 import play.api.test.Helpers.{FORBIDDEN, NOT_FOUND, OK, call}
 import services.{CompanyDetailsService, MetricsService}
@@ -33,7 +33,7 @@ import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.Future
 
-class CompanyDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSMocks with AuthFixture with CompanyDetailsFixture {
+class CompanyDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSMocks with AuthFixture with CompanyDetailsFixture with ControllerHelper{
 
   implicit val system = ActorSystem("CR")
   implicit val materializer = ActorMaterializer()
@@ -60,7 +60,11 @@ class CompanyDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSM
 
       val result = controller.retrieveCompanyDetails(registrationID)(FakeRequest())
       status(result) shouldBe OK
-      await(jsonBodyOf(result)) shouldBe Json.toJson(validCompanyDetailsResponse)
+
+      val json =  await(jsonBodyOf(result)).as[JsObject]
+      val links2 = links(json.value("links").as[JsObject])
+      json.as[JsObject] - "links" ++ links2 shouldBe Json.toJson(Some(validCompanyDetailsResponse))
+
     }
 
     "return a 404 - Not Found if the record does not exist" in new Setup {
@@ -112,9 +116,12 @@ class CompanyDetailsControllerSpec extends UnitSpec with MockitoSugar with SCRSM
 
       val request = FakeRequest().withBody(Json.toJson(validCompanyDetails))
       val result = call(controller.updateCompanyDetails(registrationID), request)
-      status(result) shouldBe OK
-      await(jsonBodyOf(result)) shouldBe Json.toJson(validCompanyDetailsResponse)
-    }
+
+      val json =  await(jsonBodyOf(result)).as[JsObject]
+      val links2 = links(json.value("links").as[JsObject])
+      json.as[JsObject] - "links" ++ links2 shouldBe Json.toJson(Some(validCompanyDetailsResponse))
+
+      status(result) shouldBe OK}
 
     "return a 404 - Not Found if the recorde to update does not exist" in new Setup {
       AuthenticationMocks.getCurrentAuthority(Some(validAuthority))
